% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcScaling.R
\name{calcScaling}
\alias{calcScaling}
\title{calcScaling}
\usage{
calcScaling(gdx, file = NULL, magnitude = 2)
}
\arguments{
\item{gdx}{path to a gdx file}

\item{file}{A file name the scaling GAMS code should be written to. If NULL
the code is returned by the function}

\item{magnitude}{The order of magnitude for which variables should be
scaled. All variables with average absolute values which are either below
10e(-magnitude) or above 10e(magnitude) will be scaled.}
}
\value{
A vector with the scaling GAMS code if file=NULL, otherwise nothing
is returned.
}
\description{
This function creates a GAMS file with scaling of variables. The scaling is
calculated based on a gdx file containing all variables of a run.
}
\examples{
\dontrun{
calcScaling("fulldata.gdx")
}
}
\seealso{
\code{\link{readGDX}}
}
\author{
Jan Philipp Dietrich
}
