% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx2-package.R
\docType{package}
\name{gdx2-package}
\alias{gdx2}
\alias{gdx2-package}
\title{gdx2: Interface package for GDX files in R}
\description{
A wrapper package for the gamstransfer package extending its functionality and allowing to read GDX files directly in R. It is emulating the basic features of the readGDX function in the gdx package but now based on gamstransfer instead of gdxrrw which served as a basis for gdx.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/gdx2}
  \item \doi{10.5281/zenodo.21256770}
  \item Report bugs at \url{https://github.com/pik-piam/gdx2/issues}
}

}
\author{
\strong{Maintainer}: Jan Philipp Dietrich \email{dietrich@pik-potsdam.de} (\href{https://orcid.org/0000-0002-4309-6431}{ORCID}) (affiliation: Potsdam Institute for Climate Impact Research)

Authors:
\itemize{
  \item Jan Philipp Dietrich \email{dietrich@pik-potsdam.de} (\href{https://orcid.org/0000-0002-4309-6431}{ORCID}) (affiliation: Potsdam Institute for Climate Impact Research)
}

}
\keyword{internal}
