% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGDX2.R
\name{readGDX2}
\alias{readGDX2}
\title{readGDX2}
\usage{
readGDX2(
  gdx,
  ...,
  format = "simplest",
  react = "warning",
  spatial = NULL,
  temporal = NULL,
  magpie_cells = TRUE
)
}
\arguments{
\item{gdx}{file name of a gdx file}

\item{...}{search strings defining the objects that should be read from gdx
file, with *-autocompletion. Can also be vectors containing more than one
search strings}

\item{format}{Output format. Five choices are currently available
\code{simple}, \code{simplest}, \code{first_found}, and \code{name}.
Instead of writing the full format name each format has
its own abbreviation as shown below.  \describe{
\item{simple (s)}{This returns a list of outputs.}
\item{simplest (st)}{Behaves like "simple" if
more than one object is returned. However, if only one object is read from
gdx file the magpie object itself is returned getting rid of the surrounding
list structure. This is the recommended format for interactive use.}
\item{first_found (f)}{This is a special format for the case that you
would like to read in exactly one object but you do not know exactly what
the name of the object is. Here, you can list all possible names of the
object and the function will return the first object of the list which is
found. This is especially useful writing read functions for gdx outputs of
models in which the names of a data object might change over time but the
function itself should work for all model versions. Having this format helps
to make your gdx-based functions backwards compatible to older versions of a
gdx file with different naming.}
\item{raw (r)}{This returns the data as it comes from
\code{gamstransfer::readGDX} without data class conversion.}
\item{name (n)}{In this case the function returns the name of all objects
found in the gdx which fit to the given search pattern and the given type
as vector.}}}

\item{react}{determines the reaction, when the object you would like to read
in does not exist. Available options are "warning" (NULL is returned and a
warning is send that the object is missing), "silent" (NULL is returned, but
no warning is given) and "error" (The function throws out an error)}

\item{spatial}{argument to determine the spatial columns in the dataframe to
be converted to a magclass object. Defaults to NULL.
See \code{\link[magclass]{as.magpie}} for more information.}

\item{temporal}{argument to determine the temporal columns in the dataframe to
be converted to a magclass object. Defaults to NULL.
See \code{\link[magclass]{as.magpie}} for more information.}

\item{magpie_cells}{(boolean) determines whether a set "j" gets special treatment
by replacing underscores in the set elements with dots. Active by default for
historical reasons. Can be ignored in most cases. Makes only a difference, if
1) GDX element depends on set "j", 2) set "j" contains underscores.}
}
\value{
The gdx objects read in the format set with the argument
\code{format}.
}
\description{
Function to read gdx files in R. It is a stripped-down reimplementation of
readGDX which is now based on magclass structures and uses
\code{gamstransfer} as basis.
}
\examples{
\dontrun{
readGDX2("bla.gdx", "blub*")
}
}
\author{
Jan Philipp Dietrich
}
